Use np.complex128 for simulating xeb circuits (#6672)
* Use np.complex128 for simulating xeb circuits

* update test

+++ cirq-core/cirq/experiments/xeb_simulation.py
@@ -107,7 +107,7 @@ def simulate_2q_xeb_circuits(
         # Need an actual object; not np.random or else multiprocessing will
         # fail to pickle the closure object:
         # https://github.com/quantumlib/Cirq/issues/3717
-        simulator = sim.Simulator(seed=np.random.RandomState())
+        simulator = sim.Simulator(seed=np.random.RandomState(), dtype=np.complex128)
     _simulate_2q_xeb_circuit = _Simulate_2q_XEB_Circuit(simulator=simulator)
 
     tasks = tuple(
